+++ packages/api/tests/endpoints/async-operations/test-endpoints-async-operations.js
@@ -36,7 +36,6 @@ const {
   setAuthorizedOAuthUsers,
   createAsyncOperationTestRecords,
 } = require('../../../lib/testUtils');
-const { buildFakeExpressResponse } = require('../utils');
 
 process.env.stackName = randomString();
 process.env.system_bucket = randomString();
@@ -225,7 +224,7 @@ test('del() returns a 401 bad request if id is not provided', async (t) => {
   t.true(fakeResponse.boom.badRequest.called);
 });
 
-test('DELETE returns a 404 if PostgreSQL and Elasticsearch async operation cannot be found', async (t) => {
+test('DELETE returns a 404 if PostgreSQL async operation cannot be found', async (t) => {
   const nonExistentAsyncOperation = fakeAsyncOperationFactory();
   const response = await request(app)
     .delete(`/asyncOperations/${nonExistentAsyncOperation.id}`)
@@ -235,76 +234,7 @@ test('DELETE returns a 404 if PostgreSQL and Elasticsearch async operation canno
   t.is(response.body.message, 'No record found');
 });
 
-test('DELETE deletes async operation successfully if it exists in PostgreSQL but not Elasticsearch', async (t) => {
-  const {
-    asyncOperationPgModel,
-    esAsyncOperationClient,
-    knex,
-  } = t.context;
-
-  const originalAsyncOperation = fakeAsyncOperationFactory();
-  const insertPgRecord = await translateApiAsyncOperationToPostgresAsyncOperation(
-    originalAsyncOperation,
-    knex
-  );
-  const id = insertPgRecord.id;
-  await asyncOperationPgModel.create(
-    knex,
-    insertPgRecord
-  );
-  t.true(
-    await asyncOperationPgModel.exists(knex, { id })
-  );
-
-  const response = await request(app)
-    .delete(`/asyncOperations/${id}`)
-    .set('Accept', 'application/json')
-    .set('Authorization', `Bearer ${jwtAuthToken}`)
-    .expect(200);
-  const { message } = response.body;
-
-  t.is(message, 'Record deleted');
-  t.false(
-    await asyncOperationPgModel.exists(knex, { id })
-  );
-  t.false(await esAsyncOperationClient.exists(
-    id
-  ));
-});
-
-test('DELETE deletes async operation successfully if it exists Elasticsearch but not PostgreSQL', async (t) => {
-  const {
-    asyncOperationPgModel,
-    esAsyncOperationClient,
-    esClient,
-    esIndex,
-    knex,
-  } = t.context;
-
-  const originalAsyncOperation = fakeAsyncOperationFactory();
-  const id = originalAsyncOperation.id;
-  await indexer.indexAsyncOperation(esClient, originalAsyncOperation, esIndex);
-  t.false(
-    await asyncOperationPgModel.exists(knex, { id })
-  );
-  t.true(
-    await esAsyncOperationClient.exists(id)
-  );
-
-  const response = await request(app)
-    .delete(`/asyncOperations/${id}`)
-    .set('Accept', 'application/json')
-    .set('Authorization', `Bearer ${jwtAuthToken}`)
-    .expect(200);
-  const { message } = response.body;
-
-  t.is(message, 'Record deleted');
-  t.false(
-    await esAsyncOperationClient.exists(id)
-  );
-});
-
-test('DELETE deletes the async operation from all data stores', async (t) => {
+test('DELETE deletes the async operation from the database', async (t) => {
   const {
     originalPgRecord,
   } = await createAsyncOperationTestRecords(t.context);
@@ -326,93 +256,4 @@ test('DELETE deletes the async operation from all data stores', async (t) => {
   const dbRecords = await t.context.asyncOperationPgModel
     .search(t.context.knex, { id });
   t.is(dbRecords.length, 0);
-  t.false(await t.context.esAsyncOperationClient.exists(
-    id
-  ));
-});
-
-test('del() does not remove from Elasticsearch if removing from PostgreSQL fails', async (t) => {
-  const {
-    originalPgRecord,
-  } = await createAsyncOperationTestRecords(t.context);
-  const { id } = originalPgRecord;
-
-  const fakeAsyncOperationPgModel = {
-    delete: () => {
-      throw new Error('PG something bad');
-    },
-    get: () => Promise.resolve(originalPgRecord),
-  };
-
-  const expressRequest = {
-    params: {
-      id,
-    },
-    testContext: {
-      knex: t.context.knex,
-      asyncOperationPgModel: fakeAsyncOperationPgModel,
-    },
-  };
-
-  const response = buildFakeExpressResponse();
-
-  await t.throwsAsync(
-    del(expressRequest, response),
-    { message: 'PG something bad' }
-  );
-
-  t.true(
-    await t.context.asyncOperationPgModel.exists(t.context.knex, {
-      id,
-    })
-  );
-  t.true(
-    await t.context.esAsyncOperationClient.exists(
-      id
-    )
-  );
-});
-
-test('del() does not remove from PostgreSQL if removing from Elasticsearch fails', async (t) => {
-  const {
-    originalPgRecord,
-  } = await createAsyncOperationTestRecords(t.context);
-  const { id } = originalPgRecord;
-
-  const fakeEsClient = {
-    initializeEsClient: () => Promise.resolve(),
-    client: {
-      delete: () => {
-        throw new Error('ES something bad');
-      },
-    },
-  };
-
-  const expressRequest = {
-    params: {
-      id,
-    },
-    testContext: {
-      knex: t.context.knex,
-      esClient: fakeEsClient,
-    },
-  };
-
-  const response = buildFakeExpressResponse();
-
-  await t.throwsAsync(
-    del(expressRequest, response),
-    { message: 'ES something bad' }
-  );
-
-  t.true(
-    await t.context.asyncOperationPgModel.exists(t.context.knex, {
-      id,
-    })
-  );
-  t.true(
-    await t.context.esAsyncOperationClient.exists(
-      id
-    )
-  );
 });